Add files via upload
The total mesh has now been created. So far we are only graphing the mesh. 
Accessing the data inside of the mesh to measure the distance between the measurement points and transducers might be tricky. We need to figure out a way to either store all of the locations in one large list, or measure bit by bit. I think storing all of the data in one massive list/array could be a good idea, that way when we use the matrix method we only have to use one matrix instead of 72