Add files via upload
Attempting to pull out the pressure at the mid line. This will allow us to place error bars here. 
Also attempting to follow the matrix method paper and not account for directionality in the transducers the way that we have previously done. Instead, we are counting the bottom and top spherical caps as just one transducer, and running the matrix method accordingly, treating the other set of transducers as reflectors. This may cause issues because of the j/... terms